+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Hide ui elements
 * [PR-7](https://github.com/itk-kimai/kimai-plugin-AarhusKommuneBundle/pull/7)
   Configuration documentation
+* [PR-6](https://github.com/itk-kimai/kimai-plugin-AarhusKommuneBundle/pull/6)
+  Improved handling of default timesheet entries
 * [PR-5](https://github.com/itk-kimai/kimai-plugin-AarhusKommuneBundle/pull/5)
   Web Accessibility Statement
 * [PR-4](https://github.com/itk-kimai/kimai-plugin-AarhusKommuneBundle/pull/4)